Create test file for Task state handler
