+++ __tests__/StateMachine.test.ts
@@ -1,6 +1,5 @@
 import type { StateMachineDefinition } from '../src/typings/StateMachineDefinition';
 import { StateMachine } from '../src/StateMachine';
-import { LambdaClient } from '../src/aws/LambdaClient';
 import * as ChoiceHelperModule from '../src/ChoiceHelper';
 import * as utilModule from '../src/util';
 
@@ -9,82 +8,6 @@ afterEach(() => {
 });
 
 describe('State Machine', () => {
-  describe('Task State', () => {
-    const mockInvokeFunction = jest.fn();
-
-    beforeEach(() => {
-      jest.spyOn(LambdaClient.prototype, 'invokeFunction').mockImplementation(mockInvokeFunction);
-    });
-
-    test('should invoke Lambda function specified in `Resource` field', async () => {
-      const definition: StateMachineDefinition = {
-        StartAt: 'TaskState',
-        States: {
-          TaskState: {
-            Type: 'Task',
-            Resource: 'mock-arn',
-            End: true,
-          },
-        },
-      };
-      const input = { input1: 'input string', input2: 10 };
-
-      const stateMachine = new StateMachine(definition, { checkArn: false });
-      await stateMachine.run(input);
-
-      expect(mockInvokeFunction).toHaveBeenCalledWith('mock-arn', { input1: 'input string', input2: 10 });
-    });
-
-    test('should return result from invoked Lambda function', async () => {
-      const definition: StateMachineDefinition = {
-        StartAt: 'TaskState',
-        States: {
-          TaskState: {
-            Type: 'Task',
-            Resource: 'mock-arn',
-            End: true,
-          },
-        },
-      };
-      const input = { num1: 5, num2: 3 };
-
-      mockInvokeFunction.mockReturnValue(input.num1 + input.num2);
-      const stateMachine = new StateMachine(definition, { checkArn: false });
-      const result = await stateMachine.run(input);
-
-      expect(mockInvokeFunction).toHaveBeenCalledWith('mock-arn', { num1: 5, num2: 3 });
-      expect(result).toBe(8);
-    });
-
-    test('should call function specified in local handler override option instead of invoking Lambda function', async () => {
-      const definition: StateMachineDefinition = {
-        StartAt: 'TaskState',
-        States: {
-          TaskState: {
-            Type: 'Task',
-            Resource: 'mock-arn',
-            End: true,
-          },
-        },
-      };
-      const input = { num1: 5, num2: 3 };
-
-      const localHandlerFn = jest.fn((event) => event.num1 + event.num2);
-      const stateMachine = new StateMachine(definition, { checkArn: false });
-      const result = await stateMachine.run(input, {
-        overrides: {
-          taskResourceLocalHandlers: {
-            TaskState: localHandlerFn,
-          },
-        },
-      });
-
-      expect(localHandlerFn).toHaveBeenCalledWith(input);
-      expect(mockInvokeFunction).not.toHaveBeenCalled();
-      expect(result).toBe(8);
-    });
-  });
-
   describe('Map State', () => {
     test('should iterate over the array referenced by `ItemsPath`', async () => {
       const definition: StateMachineDefinition = {
